Move screens to a corresponding package
