+++ core/src/se2/gruppe2/moving_maze/MovingMazeGame.java
@@ -1,14 +1,10 @@
 package se2.gruppe2.moving_maze;
 
-import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.Game;
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import se2.gruppe2.moving_maze.screens.*;
 
 public class MovingMazeGame extends Game {
