+++ core/src/se2/gruppe2/moving_maze/screens/GameScreen.java
@@ -1,22 +1,23 @@
-package se2.gruppe2.moving_maze;
+package se2.gruppe2.moving_maze.screens;
 
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.utils.ScreenUtils;
+import se2.gruppe2.moving_maze.MovingMazeGame;
 
-public class OptionScreen implements Screen {
+public class GameScreen implements Screen {
 
     final MovingMazeGame game;
     OrthographicCamera camera;
 
-    public OptionScreen(final MovingMazeGame game) {
+    public GameScreen(final MovingMazeGame game) {
         this.game = game;
         camera = MovingMazeGame.getStandardizedCamera();
     }
 
     @Override
     public void show() {
-
+        System.out.println("Gamescreen has been shown!");
     }
 
     @Override
@@ -25,7 +26,7 @@ public class OptionScreen implements Screen {
         game.batch.setProjectionMatrix(camera.combined);
 
         game.batch.begin();
-        game.font.draw(game.batch, "Option screen", 100, 100);
+        game.font.draw(game.batch, "Game screen", 100, 100);
         game.batch.end();
     }
