+++ core/src/se2/gruppe2/moving_maze/screens/RuleScreen.java
@@ -1,15 +1,16 @@
-package se2.gruppe2.moving_maze;
+package se2.gruppe2.moving_maze.screens;
 
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.utils.ScreenUtils;
+import se2.gruppe2.moving_maze.MovingMazeGame;
 
-public class CreateSessionScreen implements Screen {
+public class RuleScreen implements Screen {
 
     final MovingMazeGame game;
     OrthographicCamera camera;
 
-    public CreateSessionScreen(final MovingMazeGame game) {
+    public RuleScreen(final MovingMazeGame game) {
         this.game = game;
         camera = MovingMazeGame.getStandardizedCamera();
     }
@@ -25,7 +26,7 @@ public class CreateSessionScreen implements Screen {
         game.batch.setProjectionMatrix(camera.combined);
 
         game.batch.begin();
-        game.font.draw(game.batch, "Create session ('Lobby')", 100, 100);
+        game.font.draw(game.batch, "Rule screen", 100, 100);
         game.batch.end();
     }
